Added the CreateUsername and CreateUserId Servlet and modified Datastore Manager.

+++ src/main/java/com/google/sps/managers/DatastoreManager.java
@@ -128,12 +128,12 @@ public class DatastoreManager implements IDataManager {
     String gameDescription = (String) gameEntity.getProperty("gameDescription");
     String gameCreator = (String) gameEntity.getProperty("gameCreator");
     ArrayList<String> stages = (ArrayList<String>)gameEntity.getProperty("stages");
-    int numTimesPlayed = (int)gameEntity.getProperty("numTimesPlayed");
-    int numTimesFinished = (int)gameEntity.getProperty("numTimesFinished");
-    int numStarVotes = (int)gameEntity.getProperty("numStarVotes");
-    int totalStars = (int)gameEntity.getProperty("totalStars");
-    int numDifficultyVotes = (int)gameEntity.getProperty("numDifficultyVotes");
-    int totalDifficulty = (int)gameEntity.getProperty("totalDifficulty");
+    int numTimesPlayed = ((Long)gameEntity.getProperty("numTimesPlayed")).intValue();
+    int numTimesFinished = ((Long)gameEntity.getProperty("numTimesFinished")).intValue();
+    int numStarVotes = ((Long)gameEntity.getProperty("numStarVotes")).intValue();
+    int totalStars = ((Long)gameEntity.getProperty("totalStars")).intValue();
+    int numDifficultyVotes = ((Long)gameEntity.getProperty("numDifficultyVotes")).intValue();
+    int totalDifficulty = ((Long)gameEntity.getProperty("totalDifficulty")).intValue();
   
     Game.Builder game = new Game.Builder(gameID, gameName).setGameCreator(gameCreator);
     game.setGameDescription(gameDescription).setNumTimesPlayed(numTimesPlayed).setNumTimesFinished(numTimesFinished);
@@ -178,7 +178,7 @@ public class DatastoreManager implements IDataManager {
       return null;
     }
 
-    int stageNumber = (int) stageEntity.getProperty("stageNumber");
+    int stageNumber = ((Long) stageEntity.getProperty("stageNumber")).intValue();
     String key = (String) stageEntity.getProperty("key");
     String startingHint = (String) stageEntity.getProperty("startingHint");
     double latitude = (double) stageEntity.getProperty("latitude");
@@ -268,7 +268,7 @@ public class DatastoreManager implements IDataManager {
     catch(Exception e) {
       return null;
     }
-    int hintNumber = (int) hintEntity.getProperty("hintNumber");
+    int hintNumber = ((Long) hintEntity.getProperty("hintNumber")).intValue();
     String text = (String) hintEntity.getProperty("key");
     double latitude = (double) hintEntity.getProperty("latitude");
     double longitude = (double) hintEntity.getProperty("longitude");
@@ -295,12 +295,12 @@ public class DatastoreManager implements IDataManager {
       String gameDescription = (String) gameEntity.getProperty("gameDescription");
       String gameCreator = (String) gameEntity.getProperty("gameCreator");
       ArrayList<String> stages = (ArrayList<String>)gameEntity.getProperty("stages");
-      int numTimesPlayed = (int)gameEntity.getProperty("numTimesPlayed");
-      int numTimesFinished = (int)gameEntity.getProperty("numTimesFinished");
-      int numStarVotes = (int)gameEntity.getProperty("numStarVotes");
-      int totalStars = (int)gameEntity.getProperty("totalStars");
-      int numDifficultyVotes = (int)gameEntity.getProperty("numDifficultyVotes");
-      int totalDifficulty = (int)gameEntity.getProperty("totalDifficulty");
+      int numTimesPlayed = ((Long)gameEntity.getProperty("numTimesPlayed")).intValue();
+      int numTimesFinished = ((Long)gameEntity.getProperty("numTimesFinished")).intValue();
+      int numStarVotes = ((Long)gameEntity.getProperty("numStarVotes")).intValue();
+      int totalStars = ((Long)gameEntity.getProperty("totalStars")).intValue();
+      int numDifficultyVotes = ((Long)gameEntity.getProperty("numDifficultyVotes")).intValue();
+      int totalDifficulty = ((Long)gameEntity.getProperty("totalDifficulty")).intValue();
     
       Game.Builder game = new Game.Builder(gameID, gameName).setGameCreator(gameCreator);
       game.setGameDescription(gameDescription).setNumTimesPlayed(numTimesPlayed).setNumTimesFinished(numTimesFinished);
